LayoutTests/fast/js/sputnik/Conformance/11_Expressions/11.13_Assignment_Operators/11.13.2_Compound_Assignment/S11.13.2_A2.2_T6-expected.txt

S11.13.2_A2.2_T6

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
